capstone-slackbot/pyproject.toml
[tool.poetry]
name = "capstone-slackbot"
version = "0.1.0"
description = "Slack bot for natural language database queries using PandaAI and MCP DatabaseToolbox"
authors = ["Dirk Verstraete <dirk@example.com>"]
readme = "README.md"
packages = [{include = "capstone_slackbot", from = "."}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.0.0"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"
pandasai = "^3.0.0"
pandasai-litellm = "^0.0.1"
openai = "^1.0.0"
slack-sdk = "^3.20.0"
slack-bolt = "^1.18.0"
# MCP Python SDK - install from GitHub until available on PyPI
# mcp = { git = "https://github.com/modelcontextprotocol/python-sdk.git", branch = "main" }
# For now, MCP is optional - code handles ImportError gracefully
psycopg2-binary = "^2.9.0"
sqlalchemy = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
slack-bot = "capstone_slackbot.main:main"
mcp-server = "capstone_slackbot.mcp_server.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

